re-write test using new jws.Verifier
